Moved getRegistrationTermString to commonFunctions
